Result file names are not always the same.
The output of the rnacov command line program can change based on the
input file. Since we never know what the name is going to be the code
has to use globing to find the file it expects, based on the extension,
and use that as the target for reading the image data.