Restores download functionality.
The output file was renamed from query.out to <uploaded_file>.out, this
caused the download to fail as it was looking for the wrong file. To fix
this I copied the code from the results page that looks up the name of
the uploaded file and uses that to figure out the name of the output
file.